pyproject.toml
[project]
name = "project-2"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "google-generativeai==0.8.5",
    "pypdf2>=3.0.1",
    "python-dotenv==1.1.1",
    "streamlit==1.46.1",
]
